fix(quality): decompose the parafering / tenant / role service cluster (10 phpmd findings)
Clears all 10 remaining PHPMD architectural findings in this cluster by
extracting cohesive collaborators — no suppressions added, no baseline,
no threshold changes.

ExcessiveClassComplexity (WMC, threshold 50):
  EmailTemplateService        50 -> 22
  ZaakdossierService          50 -> 42
  RoleResolverService         51 -> 38
  MilestoneService            53 -> 18
  TenantConfigurationService  54 -> 32
  MandaatImportService        73 -> 27
  ParafeerRouteService        74 -> 45
  ParafeerActieService        79 -> 43

CouplingBetweenObjects (threshold 13):
  DsoCaseService              13 -> 12
  LibresignSigningAdapter     14 -> 12

Repo-wide phpmd: 25 -> 15 findings; the 15 that remain are byte-identical
to the pre-existing ones. RoleResolverService's class-level
CouplingBetweenObjects suppression is now unnecessary (CBO 12) and was
removed. The suppressions on the two Parafeer services were re-checked
and are still required (CBO 21 and 15), so they stay, now annotated with
the measured post-split value.

Behaviour is unchanged: every public signature is preserved via
delegation, and moved constants are aliased back on their original class.
Tests wire REAL collaborators (not mocks) so the production logic they
assert is still the logic under test.
